[Refactor] Overhaul of scheduler component (#1169)
This is a significant overhaul to Nativelink's scheduler
component. This new scheduler design is to enable a distributed
scheduling system.

The new components & definitions:
* AwaitedActionDb - An interface that is easier to work with when
  dealing with key-value storage systems.
* MemoryAwaitedActionDb - An in-memory set of hashmaps & btrees used
  to satisfy the requirements of AwaitedActionDb interface.
* ClientStateManager - A minimal interface required to satisfy the
  requirements of a client-facing scheduler.
* WorkerStateManager - A minimal interface required to satisfy the
  requirements of a worker-facing scheduler.
* MatchingEngineStateManager - A minimal interface required to
  satisfy a engine that matches queued jobs to workers.
* SimpleSchedulerStateManager - An implements that satisfies
  ClientStateManager, WorkerStateManager & MatchingEngineStateManager
  with all the logic of the previous "SimpleScheduler" logic moved
  behind each interface.
* ApiWorkerScheduler - A component that handles all knowledge about
  workers state and implmenets the WorkerScheduler interface and
  translates them into the WorkerStateManager interface.
* SimpleScheduler - Translation calls of the ClientScheduler
  interface into ClientStateManager & MatchingEngineStateManager.
  This component is currently always forwards calls to
  SimpleSchedulerStateManager then to MemoryAwaitedActionDb.
  Future changes will make these inner components dynamic via config.

In addition we have hardened the interactions of different kind of
IDs in NativeLink. Most relevant is the separation & introduction of:
* OperationId - Represents an individual  operation being requested
  to be executed that is unique across all of time.
* ClientOperationId - An ID issued to the client when the client
  requests to execute a job. This ID will point to an OperationId
  internally, but the client is never exposed to the OperationId.
* AwaitedActionHashKey - A key used to uniquely identify an action
  that is not unique across time. This means that this key might
  have multiple OperationId's that have executed it across different
  points in time. This key is used as a "fingerprint" of an operation
  that the client wants to execute and the scheduler may decide to
  join the stream onto an existing operation if this key has a hit.

Overall these changes pave the way for more robust scheduler
implementations, most notably, distributed scheduler implementations
will be easier to implement and will be introduced in followup PRs.

This commit was developed on a side branch and consisted of the
following commits with corresponding code reviews:
54ed73cc6fcaf9236a544e45a90b9071b5df6ad8
    Add scheduler metrics back (#1171)
50fdbd73983dfdbd2620ac6bfcd5c7b642b5705e
    fix formatting (#1170)
89262366544d2310e0e9a5bd91c99297112d4e27
    Merge in main and format (#1168)
9c2c7b9d3a2bfd71615fe46893ac937e6b23974c
    key as u64 (#1166)
0192051fca6e3509a6b2953ab53f77c12b4c7092
    Cleanup unused code and comments (#1165)
080df5d0c559055d12ac53275024d32e35077675
    Add versioning to AwaitedAction (#1163)
73c19c41e31fef55a3c6e81beb19a7233f0c2e8d
    Fix sequence bug in new memory store manager (#1162)
6e50d2c16fad92f2f8fb856b199f7798c30d5967
    New AwaitedActionDb implementation (#1157)
18db991e3a414c3f95c2dd5a14bfafce80358de9
    Fix test on running_actions_manager_test (#1141)
e50ef3c27d16fc840fd63b917f39f1c317f48398
    Rename workers to `worker_scheduler`
1fdd5056f357124d6aff63fe145821b795cf9780
    SimpleScheduler now uses config for action pruning (#1137)
eaaa872bc62a952fb926214a6fb11bd4ffd25bae
    Change encoding for items that are cachable (#1136)
d64705627c045d108340546a3ea135edb984163a
    Errors are now properly handles in subscription (#1135)
7c3e730e30e96fb5d08e4d43cca8d27593bcea03
    Restructure files to be more appropriate (#1131)
5e98ec9f67f56744376e060b6aa423f1ff401a47
    ClientAwaitedAction now uses a channel to notify drops happened (#1130)
52beaf9c727cd396ce4b5b4095bc8dc7be8ce940
    Cleanup unused structs (#1128)
e86fe0826b3512e78804744a8f1db2ec4c5b21ea
    Remove all uses of salt and put under ActionUniqueQualifier (#1126)
3b860367ba131c918127b6426e791b9ca0ca8b6b
    Remove all need for workers to know about ActionId (#1125)
5482d7f5e5f84c2ea80d79c703624c73b6ee36cc
    Fix bazel build and test on dev (#1123)
ba52c7fb76bba56ff5485b91cd516712f867dc64
    Implement get_action_info to all ActionStateResult impls (#1118)
2fa4fee48374ed0cf12dcd88225d6d208c0eeeb2
    Remove MatchingEngineStateManager::remove_operation (#1119)
34dea0633f8b5198d9c51904beee8f4c617fc822
    Remove unused proto field (#1117)
3070a40935c1a490b530af8551329c44bb4500ff
    Remove metrics from new scheduler (#1116)
e95adfcd6035d3cecac74e67829b83e44a598228
    StateManager will now cleanup actions on client disconnect (#1107)
6f8c001d1a96a0dc74a5c22716ba2f54e4c9f1b1
    Fix worker execution issues (#1114)
d353c3003437609547c5dcc8738637b13151a9b5
    rename set_priority to upgrade_priority (#1112)
0d93671ba51be41a1a6b0297260cfc5ca0766ebc
    StateManager can now be notified of noone listeneing (#1093)
cfc0cf62c673dc399938247b0c3abe55863c2687
    ActionScheduler will now use ActionListener instead of tokio::watch (#1091)
d70d31d74cdc337861ebddd8d19201b088b8faa6
    QA fixes for scheduler-v2 (#1092)
f2cea0cc3fcc259b9a4b7257f2eae807d7e99d05
    [Refactor] Complete rewrite of SimpleScheduler
34d93b7d1cda7c0f4a430364b7588002ab612494
    [Refactor] Move worker notification in SimpleScheduler under Workers
b9d970264b59201e13053844a277092164c8416a
    [Refactor] Moves worker logic back to SimpleScheduler
7a16e2e6043b17e7813e41450b4de9c40de435f4
    [Refactor] Move scheduler state behind mute

+++ nativelink-config/src/schedulers.rs
@@ -98,7 +98,7 @@ pub struct SimpleScheduler {
     /// a WaitExecution is called after the action has completed.
     /// Default: 60 (seconds)
     #[serde(default, deserialize_with = "convert_duration_with_shellexpand")]
-    pub retain_completed_for_s: u64,
+    pub retain_completed_for_s: u32,
 
     /// Remove workers from pool once the worker has not responded in this
     /// amount of time in seconds.
